requirements.txt
web3~=5.31.3
aiohttp~=3.8.4
python-dotenv~=1.0.0
